rename/#54 : SuwikiAddClassAppBar, SuwikiAppBar -> SuwikiAppBarWithTextBtn, SuwikiAppBarWithTitle
